I changed the mail sender IP we are using when we run spf.query.
We were using 127.0.0.1.  When the domain being scanned uses the ptr
mechanism, a reverse DNS lookup is done against the IP of the mail
sender.  This fails for 127.0.0.1.  The solution is to use an IP that
1. Has a valid DNS PTR record
2. Is not used as a mail server by anyone

I tried using c1b1.ncats.cyber.dhs.gov, since that has a valid DNS PTR
record in our AWS Route 53 setup and it is not used as a mail server.
I found, though, that Google DNS does not return the PTR record for
that hostname.  Until that is resolved I'm using an IP that is used by
www.google.com that does have a valid PTR record and should not used
as a mail server by anyone we scan.