Fixed a bug where domains were being flagged as "not live" when a lookup of the DMARC record returned NXDOMAIN.
Dave Redmin noticed that a lot of domains were being incorrectly
flagged as "not live".  The problem turned out to that domains are
being flagged as "not live" when a lookup of the DMARC record returns
NXDOMAIN.  This is incorrect, since domain.com may well exist even
when _dmarc.domain.com does not.