Fixing a bug with the public suffix list (PSL) caching
Previously, if the PSL file did not exist and had to be downloaded,
the stream object was read into the cache file and then passed to
publicsuffix.PublicSuffixList().  At this point the stream object was
empty, so the code was effectively using an empty PSL until get_psl()
was called again and the PSL was read in from the disk cache.

This change fixes that by modifying get_psl().  Now that function
checks to see if the cache file exists and is less than 24 hours old
and downloads it if either of those checks fails.  Then it opens the
cache file and hands that file stream (which is guaranteed to be
unread and full of data) off to publicsuffix.PublicSuffixList().