Stop looking up the SPF record twice
If we include the SPF record (as obtained by following redirects) as
an argument to spf.query.check(), then that method will not look up
the SPF record a second time.  I suspect that these repeated DNS
queries may be occasionally failing, and that is what is responsible
for the "Result unexpectedly differs" errors that are seen in the
trustymail results.