Change the way we validate the SPF record
Now pyspf does all the work for us, and we don't try to check anything
manually.  We were doing the manual check incorrectly before, anyway,
since if an SPF record has an include directive then the mechanism
from the included record is what counts, NOT the mechanism from the
initial SPF record for the domain being evaluated.