Force smtplib to only use IPv4 addresses
By default, smtplib looks for A and AAAA records from DNS and uses the
first one that it can connect to.  What I find when running in
Lambda (at least in my VPC that doesn't support IPv6) is that when DNS
returns IPv6 an address I get a low level "errno 97 - Address family
not supported by protocol" error and the other addresses returned by
DNS are not even attempted.  Therefore the hostname is not scanned at
all.

To get around this I look up the A record and use that instead of the
hostname in DNS when I call smtp_connection.connect().