Resolve #119 by adjusting run script and Dockerfile
It appears that the bind volume was what caused an issue with the run script.
I adjusted the run script to mount the current directory to a different
directory in the image, and then set the working directory for docker run to
that directory. This way you can still use local files for trustymail
execution. The docker build command will still update the image if changes are
detected.