Add flake8 configuration file
After blackening the repo I saw that the Travis build was failing. I saw that
it was due to incompatibility between default the default flake8 configuration
and the newly blackened code. So I added the .flake8 configuration file we use
in skeleton-python-library as an interim to fully switching this project over
to GitHub Actions from Travis.